+++ tests/testthat/test-graphics.R
@@ -137,3 +137,14 @@ test_that("Rplots.pdf files are not created", {
   evaluate(file("plot.R"))
   expect_false(file.exists("Rplots.pdf"))
 })
+
+# https://github.com/yihui/knitr/issues/2297
+test_that("existing plots will not leak into evaluate()", {
+  pdf(NULL)
+  dev.control('enable')
+  d <- dev.cur()
+  plot(1, 1)
+  ev <- evaluate(c('dev.new()', 'dev.off()', 'plot.new()', 'plot(1:10, 1:10)'))
+  dev.off(d)
+  expect_equal(tail(classes(ev), 6), c('source', 'character', 'recordedplot')[c(1, 2, 1, 3, 1, 3)])
+})